[#52] Refactor components
